prov/tcp: Initial support of io_uring for the TCP provider
This patch implements the support of io_uring in the tcp provider.
The patch is rather small and it doesn't impact much the core of
the TCP provider, which should limit the regressions. In addition,
the patch allows for turning on/off io_uring support at runtime
with the environment variable FI_TCP_URING.

The availability of io_uring is automatically detected by the
configure script. By default io_uring is disabled even if libfabric was
compiled with its support. To enable io_uring at runtime, one needs
to export FI_TCP_URING=1

In order to avoid costly locking mechanisms, each endpoint has its own
io_uring and 4 associated contexts: 2 for TX/RX operations and 2
additional for canceled TX/RS operations.

Each context keeps a variable that reflects the state of the context:
- IDLE: the context is idle and can be used to submit an async operation
- BUSY: the context has a pending async operation waiting for completion.
        In this state, the context doesn't accept new operations and the
        caller must poll the EP ring to progress the corresponding IO operation.
- DONE: the async operation associated to the context has completed and
        it'ss waiting for processing.

Socket operations with io_uring happen as follows:
- If the context is IDLE, a new SQE entry is filled in and submitted to the
  EP ring. The context is flagged as BUSY and the operation returns EAGAIN
  to the caller so it can later retry the operation to check for completion.
- The CQ progress function polls the EP ring and mark completed contexts
  as 'DONE' and store the result of the operation (error or number of bytes
  read/written) is stored in the context.
- When the socket operation is retried and if the context is 'DONE', the
  result of the operation is given to the caller, which can proceed.

The TCP provider uses an epoll set of file descriptors to progress
the endpoints associated to a given completion queue. In addition to
the socket FD, we register both TX and RX ring FDs for POLLIN events in
the epoll set so we get notified when a CQE is available.

Finally, the patch modifies tcpx_update_epoll() so we don't always
wait for the socket events if a SQE event is waiting for completion,
which would wake up epoll when it would complete.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>