bsock: Add handling of io_uring error codes
io_uring operations in the sockapi interface would return -FI_EIOURING_PREP
if the operation was prepared in the ring and is waiting for submission.

recv() and recvv() require some changes to the bsock API to return how
many bytes were read from the byteq to the caller (similar to what was done
for zerocopy send). I added a couple of assertion failures to detect
unimplemented features related to io_uring.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>