prov/verbs: Allow for large TX queues with limited (or no) inline data
Using large TX queues with the verbs provider would cause fi_getinfo()
to return an empty list of verbs adapters because the call to
ibv_create_qp() executed as part of fi_getinfo() would fail with EINVAL.

The failure happens because the code allocates the QP with the maximum
amount of inline data supported by the adapter, which is empirically
determined by vrb_find_max_inline(). The problem is that using inline
data limits the TX queue size that can be allocated.

The fix implemented in this patch is to set max_inline_data = 0 when
the QP is created, then update info->tx_attr->inject_size with the value
returned by vrb_find_max_inline() after the QP is created. The code in
vrb_find_max_inline() guarantees that the calculated inline value is
correct as it is also tested with a fake QP creation.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>